Be a bit more soft in the edges
Eg. The moment package uses `require(“lang” + k);` our parser says
`library “lang” not found on the server - but on the client we have to
include the dependency “lang” - even if never used. If the user tries
to use it we do show a warning on the client too.
I’m not sure if we should actually throw an error - the user is doing
something wrong.