First step towards library globals WIP
So this commit rigs the parser to detect a library global like:
```js
  // I’m using the library without any require statements
  // I simply expects famono to figure out dependencies and adding those
  // to my code - sweet pattern
  famous.core.Surface();
```

So only the parsing part is up, it prints out dependencies when testing
it and it looks promising.

Next is having a proper way of preparing the external libraries for
this new pattern, this will require a better js parser. Thinking
further it could be a good idea to initialize the use of something like
esprima eg. for making fine-grained client dependencies…

We should also create an updated version of a library dependency tree
having it more fine-grained.
The library code should be inserted pre meteor so library code have to
be initialized by the app / package code. This pattern makes a lot of
sense and makes stuff much easier to work with. So library code is more
static where packages can do stuff.

The parse code now returns an array of library globals e.g.:
`globals: [ { library: 'famous', dependency: 'famous.test' } ]`

Really looking forwards to this